+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed `GetTimeOffs()` date argument format (YYYY-MM-DD)
+
 ## [0.0.1] - 2022-12-08
 
 - Add `GetEmployees()` and `GetEmployee(id)` to handle `GET /company/employees`