+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2023-02-22
+
 ### Fixed
 
 - Fixed `GetEmployees()` only returning first page of data
@@ -27,5 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Authenticate()` to handle `POST /auth`
 - Add Personio API v1 client
 
-[Unreleased]: https://github.com/giantswarm/personio-go/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/giantswarm/personio-go/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/giantswarm/personio-go/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/giantswarm/personio-go/releases/tag/v0.0.1